Refactor domain models into Aggregates namespace
Organized domain models structure
