+++ MSN.Domain/Domains/Catalog/Aggregates/OptionDefinition.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MSN.Domain.Domains.Catalog
+namespace MSN.Domain.Domains.Catalog.Aggregates
 {
     internal class OptionDefinition
     {
