+++ MSN.Domain/Domains/Catalog/Aggregates/ProductVariant.cs
@@ -1,6 +1,6 @@
 using MSN.Framework.BaseModel;
 
-namespace MSN.Domain.Domains.Catalog
+namespace MSN.Domain.Domains.Catalog.Aggregates
 {
     public class ProductVariant : BaseModel
     {
